Passing tests for guarded routes - rooom for improvement
