+++ tests/Browser/Pages/Dashboard.php
@@ -13,7 +13,7 @@ class Dashboard extends Page
      */
     public function url()
     {
-        return '/';
+        return '/login';
     }
 
     /**
